// src/app/Helper/models/SubCategory.ts
export class SubCategory {
    ID: number;
    Name: string;
    Series: string;
    Category: string;
    Department: string;
    Description: string;
    CreatedAt?:Date;
    UpdatedAt?: Date;
    CreatedByUserID?: number;
    IsActive?: boolean;
    ClassificationID?: number;
    DepartmentID?: number;
    CategoryID?: number;
    SeriesID?: number;
    Classification: string;
    UpdatePermission: string;
    
}